+++ java-io-base/src/main/java/internal/io/FunctionalFileFormatter.java
@@ -7,14 +7,15 @@ import nbbrd.io.function.IOBiConsumer;
 import java.io.IOException;
 import java.io.OutputStream;
 
+import static nbbrd.io.Resource.uncloseableOutputStream;
+
 @lombok.RequiredArgsConstructor
 public final class FunctionalFileFormatter<T> implements FileFormatter<T> {
 
-    @NonNull
-    private final IOBiConsumer<? super T, ? super OutputStream> function;
+    private final @NonNull IOBiConsumer<? super T, ? super OutputStream> function;
 
     @Override
     public void formatStream(@NonNull T value, @NonNull OutputStream resource) throws IOException {
-        function.acceptWithIO(value, resource);
+        function.acceptWithIO(value, uncloseableOutputStream(resource));
     }
 }