+++ java-io-base/src/main/java/internal/io/text/LegacyFiles.java
@@ -17,26 +17,17 @@
 package internal.io.text;
 
 import lombok.NonNull;
-import nbbrd.io.Resource;
 import nbbrd.io.function.IOSupplier;
 
 import java.io.*;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.file.AccessDeniedException;
-import java.nio.file.FileSystemException;
-import java.nio.file.NoSuchFileException;
 
 /**
  * @author Philippe Charles
  */
 @lombok.experimental.UtilityClass
 public class LegacyFiles {
-
-    public static InputStream openResource(Class<?> anchor, String name) throws IOException {
-        return Resource.getResourceAsStream(anchor, name)
-                .orElseThrow(() -> new IOException("Missing resource '" + name + "' of '" + anchor.getName() + "'"));
-    }
 
     public static Reader openReader(CharSequence source) {
         return new StringReader(source.toString());
@@ -52,7 +43,7 @@ public class LegacyFiles {
 
     @NonNull
     public static InputStream openInputStream(@NonNull File source) throws IOException {
-        return new BufferedInputStreamWithFile(checkSource(source));
+        return new BufferedInputStreamWithFile(FileSystemExceptions.checkSource(source));
     }
 
     public static Writer openWriter(IOSupplier<? extends Writer> source) throws IOException {
@@ -65,32 +56,7 @@ public class LegacyFiles {
 
     @NonNull
     public static OutputStream openOutputStream(@NonNull File target) throws IOException {
-        return new BufferedOutputStream(new FileOutputStream(checkTarget(target)));
-    }
-
-    public static File checkSource(@NonNull File source) throws FileSystemException {
-        checkExist(source);
-        checkIsFile(source);
-        return source;
-    }
-
-    public static File checkTarget(@NonNull File target) throws FileSystemException {
-        if (target.exists()) {
-            checkIsFile(target);
-        }
-        return target;
-    }
-
-    public static void checkExist(@NonNull File source) throws FileSystemException {
-        if (!source.exists()) {
-            throw new NoSuchFileException(source.getPath());
-        }
-    }
-
-    public static void checkIsFile(@NonNull File source) throws FileSystemException {
-        if (!source.isFile()) {
-            throw new AccessDeniedException(source.getPath());
-        }
+        return new BufferedOutputStream(new FileOutputStream(FileSystemExceptions.checkTarget(target)));
     }
 
     public String toSystemId(@NonNull File file) {