+++ java-io-base/src/test/java/_test/io/ResourceId.java
@@ -2,17 +2,15 @@ package _test.io;
 
 import internal.io.text.InternalTextResource;
 import lombok.NonNull;
-import nbbrd.io.Resource;
-import nbbrd.io.text.TextResource;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.Reader;
+import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+
+import static nbbrd.io.Resource.newInputStream;
+import static nbbrd.io.text.TextResource.newBufferedReader;
 
 @lombok.Value
 public class ResourceId {
@@ -24,7 +22,7 @@ public class ResourceId {
     String name;
 
     public InputStream open() throws IOException {
-        return Resource.getResourceAsStream(anchor, name).orElseThrow(IOException::new);
+        return newInputStream(anchor, name);
     }
 
     public Path copyTo(Path temp) throws IOException {
@@ -35,8 +33,20 @@ public class ResourceId {
         return result;
     }
 
+    public byte[] toBytes() throws IOException {
+        ByteArrayOutputStream result = new ByteArrayOutputStream();
+        try (InputStream input = open()) {
+            byte[] buffer = new byte[8192];
+            int read;
+            while ((read = input.read(buffer, 0, buffer.length)) >= 0) {
+                result.write(buffer, 0, read);
+            }
+        }
+        return result.toByteArray();
+    }
+
     public BufferedReader open(Charset encoding) throws IOException {
-        return TextResource.getResourceAsBufferedReader(anchor, name, encoding).orElseThrow(IOException::new);
+        return newBufferedReader(anchor, name, encoding.newDecoder());
     }
 
     public String copyToString(Charset encoding) throws IOException {