+++ java-io-base/src/test/java/internal/io/text/LegacyFilesTest.java
@@ -16,86 +16,10 @@
  */
 package internal.io.text;
 
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.AccessDeniedException;
-import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
-import java.nio.file.Path;
-
-import static org.assertj.core.api.Assertions.*;
-
 /**
  * @author Philippe Charles
  */
 public class LegacyFilesTest {
 
-    @Test
-    public void testCheckExist(@TempDir Path temp) throws IOException {
-        assertThatNullPointerException()
-                .isThrownBy(() -> LegacyFiles.checkExist(null));
 
-        {
-            File file = Files.createFile(temp.resolve("exampleFile")).toFile();
-
-            assertThatCode(() -> LegacyFiles.checkExist(file))
-                    .doesNotThrowAnyException();
-
-            file.delete();
-            assertThatIOException()
-                    .isThrownBy(() -> LegacyFiles.checkExist(file))
-                    .isInstanceOf(NoSuchFileException.class)
-                    .withMessageContaining(file.getPath());
-        }
-
-        {
-            File folder = Files.createDirectory(temp.resolve("exampleDir")).toFile();
-
-            assertThatCode(() -> LegacyFiles.checkExist(folder))
-                    .doesNotThrowAnyException();
-
-            folder.delete();
-            assertThatIOException()
-                    .isThrownBy(() -> LegacyFiles.checkExist(folder))
-                    .isInstanceOf(NoSuchFileException.class)
-                    .withMessageContaining(folder.getPath());
-        }
-    }
-
-    @Test
-    public void testCheckIsFile(@TempDir Path temp) throws IOException {
-        assertThatNullPointerException()
-                .isThrownBy(() -> LegacyFiles.checkIsFile(null));
-
-        {
-            File file = Files.createFile(temp.resolve("exampleFile")).toFile();
-
-            assertThatCode(() -> LegacyFiles.checkIsFile(file))
-                    .doesNotThrowAnyException();
-
-            file.delete();
-            assertThatIOException()
-                    .isThrownBy(() -> LegacyFiles.checkIsFile(file))
-                    .isInstanceOf(AccessDeniedException.class)
-                    .withMessageContaining(file.getPath());
-        }
-
-        {
-            File folder = Files.createDirectory(temp.resolve("exampleDir")).toFile();
-
-            assertThatIOException()
-                    .isThrownBy(() -> LegacyFiles.checkIsFile(folder))
-                    .isInstanceOf(AccessDeniedException.class)
-                    .withMessageContaining(folder.getPath());
-
-            folder.delete();
-            assertThatIOException()
-                    .isThrownBy(() -> LegacyFiles.checkIsFile(folder))
-                    .isInstanceOf(AccessDeniedException.class)
-                    .withMessageContaining(folder.getPath());
-        }
-    }
 }