+++ java-io-base/src/test/java/nbbrd/io/text/TextResourceTest.java
@@ -19,11 +19,11 @@ package nbbrd.io.text;
 import org.junit.jupiter.api.Test;
 
 import java.io.*;
+import java.nio.charset.CharsetDecoder;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static nbbrd.io.text.TextResource.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatNullPointerException;
+import static org.assertj.core.api.Assertions.*;
 
 /**
  * @author Philippe Charles
@@ -31,18 +31,38 @@ import static org.assertj.core.api.Assertions.assertThatNullPointerException;
 public class TextResourceTest {
 
     @Test
-    @SuppressWarnings("null")
+    @SuppressWarnings({"null", "DataFlowIssue", "deprecation"})
     public void testGetResourceAsBufferedReader() throws IOException {
         assertThatNullPointerException().isThrownBy(() -> getResourceAsBufferedReader(null, "", UTF_8));
         assertThatNullPointerException().isThrownBy(() -> getResourceAsBufferedReader(TextResourceTest.class, null, UTF_8));
         assertThatNullPointerException().isThrownBy(() -> getResourceAsBufferedReader(TextResourceTest.class, "", null));
 
         assertThat(getResourceAsBufferedReader(TextResourceTest.class, "missing", UTF_8)).isEmpty();
-        try (BufferedReader reader = getResourceAsBufferedReader(TextResourceTest.class, "/nbbrd/io/text/hello.txt", UTF_8).get()) {
+        try (BufferedReader reader = getResourceAsBufferedReader(TextResourceTest.class, "/nbbrd/io/text/hello.txt", UTF_8).orElseThrow(IOException::new)) {
             assertThat(reader.lines()).contains("world");
         }
     }
 
+    @Test
+    @SuppressWarnings({"null", "resource", "DataFlowIssue"})
+    public void testNewBufferedReaderOfResource() throws IOException {
+        CharsetDecoder utf8 = UTF_8.newDecoder();
+
+        assertThatNullPointerException().isThrownBy(() -> newBufferedReader(null, "", utf8));
+        assertThatNullPointerException().isThrownBy(() -> newBufferedReader(TextResourceTest.class, null, utf8));
+        assertThatNullPointerException().isThrownBy(() -> newBufferedReader(TextResourceTest.class, "", null));
+
+        assertThatIOException()
+                .isThrownBy(() -> newBufferedReader(TextResourceTest.class, "missing", utf8))
+                .withMessageContaining("missing")
+                .withMessageContaining(TextResourceTest.class.getName());
+
+        try (BufferedReader reader = newBufferedReader(TextResourceTest.class, "/nbbrd/io/text/hello.txt", utf8)) {
+            assertThat(reader.lines()).contains("world");
+        }
+    }
+
+    @SuppressWarnings({"resource", "DataFlowIssue"})
     @Test
     public void testNewBufferedReader() throws IOException {
         byte[] bytes = "world".getBytes(UTF_8);
@@ -55,6 +75,7 @@ public class TextResourceTest {
         }
     }
 
+    @SuppressWarnings({"resource", "DataFlowIssue"})
     @Test
     public void testNewBufferedWriter() throws IOException {
         assertThatNullPointerException().isThrownBy(() -> newBufferedWriter(null, UTF_8.newEncoder()));