+++ java-io-xml-bind/src/main/java/nbbrd/io/xml/bind/Jaxb.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.io.xml.bind;
 
+import internal.io.text.FileSystemExceptions;
 import internal.io.text.LegacyFiles;
 import lombok.NonNull;
 import nbbrd.design.StaticFactoryMethod;
@@ -33,6 +34,8 @@ import javax.xml.stream.XMLStreamReader;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+
+import static nbbrd.io.Resource.uncloseableInputStream;
 
 /**
  * @author Philippe Charles
@@ -103,7 +106,7 @@ public class Jaxb {
 
         @Override
         public @NonNull T parseFile(@NonNull File source) throws IOException {
-            LegacyFiles.checkSource(source);
+            FileSystemExceptions.checkSource(source);
             Unmarshaller engine = factory.getWithIO();
 
             return cast(!ignoreXXE
@@ -113,7 +116,7 @@ public class Jaxb {
 
         @Override
         public @NonNull T parseFile(@NonNull File source, @NonNull Charset encoding) throws IOException {
-            LegacyFiles.checkSource(source);
+            FileSystemExceptions.checkSource(source);
             Unmarshaller engine = factory.getWithIO();
 
             return cast(!ignoreXXE
@@ -135,8 +138,8 @@ public class Jaxb {
             Unmarshaller engine = factory.getWithIO();
 
             return cast(!ignoreXXE
-                    ? parseStreamXXE(engine, resource, xxeFactory.getWithIO())
-                    : parseStream(engine, resource));
+                    ? parseStreamXXE(engine, uncloseableInputStream(resource), xxeFactory.getWithIO())
+                    : parseStream(engine, uncloseableInputStream(resource)));
         }
 
         @Override
@@ -144,8 +147,8 @@ public class Jaxb {
             Unmarshaller engine = factory.getWithIO();
 
             return cast(!ignoreXXE
-                    ? parseStreamXXE(engine, resource, xxeFactory.getWithIO())
-                    : parseStream(engine, resource));
+                    ? parseStreamXXE(engine, uncloseableInputStream(resource), xxeFactory.getWithIO())
+                    : parseStream(engine, uncloseableInputStream(resource)));
         }
 
         @SuppressWarnings("unchecked")
@@ -272,7 +275,7 @@ public class Jaxb {
 
         @Override
         public void formatFile(@NonNull T value, @NonNull File target) throws IOException {
-            LegacyFiles.checkTarget(target);
+            FileSystemExceptions.checkTarget(target);
             try {
                 getEngine(getDefaultEncoding()).marshal(value, target);
             } catch (JAXBException ex) {