+++ java-io-xml/src/main/java/nbbrd/io/xml/Sax.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.io.xml;
 
+import internal.io.text.FileSystemExceptions;
 import internal.io.text.LegacyFiles;
 import lombok.NonNull;
 import nbbrd.design.StaticFactoryMethod;
@@ -34,6 +35,8 @@ import java.util.Locale;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
+
+import static nbbrd.io.Resource.uncloseableInputStream;
 
 /**
  * @author Philippe Charles
@@ -148,12 +151,12 @@ public class Sax {
 
         @Override
         public @NonNull T parseFile(@NonNull File source) throws IOException {
-            return doParse(newInputSource(LegacyFiles.checkSource(source)));
+            return doParse(newInputSource(FileSystemExceptions.checkSource(source)));
         }
 
         @Override
         public @NonNull T parseFile(@NonNull File source, @NonNull Charset encoding) throws IOException {
-            return doParse(newInputSource(LegacyFiles.checkSource(source), encoding));
+            return doParse(newInputSource(FileSystemExceptions.checkSource(source), encoding));
         }
 
         @Override
@@ -163,12 +166,12 @@ public class Sax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParse(new InputSource(resource));
+            return doParse(new InputSource(uncloseableInputStream(resource)));
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            InputSource input = new InputSource(resource);
+            InputSource input = new InputSource(uncloseableInputStream(resource));
             input.setEncoding(encoding.name());
             return doParse(input);
         }