+++ java-io-xml/src/main/java/nbbrd/io/xml/Stax.java
@@ -34,6 +34,8 @@ import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
+import static nbbrd.io.Resource.uncloseableInputStream;
+
 /**
  * @author Philippe Charles
  */
@@ -173,12 +175,12 @@ public class Stax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParseOrClose(o -> o.createXMLStreamReader(resource), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLStreamReader(uncloseableInputStream(resource)), NOTHING_TO_CLOSE);
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            return doParseOrClose(o -> o.createXMLStreamReader(resource, encoding.name()), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLStreamReader(uncloseableInputStream(resource), encoding.name()), NOTHING_TO_CLOSE);
         }
 
         @NonNull
@@ -274,12 +276,12 @@ public class Stax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParseOrClose(o -> o.createXMLEventReader(resource), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLEventReader(uncloseableInputStream(resource)), NOTHING_TO_CLOSE);
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            return doParseOrClose(o -> o.createXMLEventReader(resource, encoding.name()), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLEventReader(uncloseableInputStream(resource), encoding.name()), NOTHING_TO_CLOSE);
         }
 
         private T doParseOrClose(StaxFunction<XMLInputFactory, XMLEventReader> supplier, Closeable onClose) throws IOException {